Avoid overwriting modified DNS upon restore
In openvpn_disconnect(), "modify_dns revert_to_backup" is called to
restore DNS to the value backed-up by openvpn_connect(). However,
DNS might change after the backup and before a --disconnect is issued.
In that case, --disconnect causes a needless overwrite of DNS with an
old value, that might as well break the internet connection.

Use case:

  * Client connects to protonvpn.
  * Client changes access point. Internet connection breaks.
  * Client issues a --disconnect to restore internet connection.
  * Internet connection remains broken due to DNS overwrite.
    New --connect attempts will fail.